+++ libraries/AD9833/AD9833.cpp
@@ -2,23 +2,27 @@
 //    FILE: AD9833.cpp
 //  AUTHOR: Rob Tillaart
 // PURPOSE: Arduino library for AD9833 function generator
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 //     URL: https://github.com/RobTillaart/AD9833
 //
 
 #include "AD9833.h"
 
+//  FREQUENCY REGISTER BITS
+#define AD9833_FREG1        0x8000
+#define AD9833_FREG0        0x4000
+
 //  CONTROL REGISTER BITS
-#define AD9833_B28      (1 << 13)
-#define AD9833_HLB      (1 << 12)
-#define AD9833_FSELECT  (1 << 11)
-#define AD9833_PSELECT  (1 << 10)
-#define AD9833_RESET    (1 << 8)
-#define AD9833_SLEEP1   (1 << 7)
-#define AD9833_SLEEP12  (1 << 6)
-#define AD9833_OPBITEN  (1 << 5)
-#define AD9833_DIV2     (1 << 3)
-#define AD9833_MODE     (1 << 1)
+#define AD9833_B28          (1 << 13)
+#define AD9833_HLB          (1 << 12)
+#define AD9833_FSELECT      (1 << 11)
+#define AD9833_PSELECT      (1 << 10)
+#define AD9833_RESET        (1 << 8)
+#define AD9833_SLEEP1       (1 << 7)
+#define AD9833_SLEEP12      (1 << 6)
+#define AD9833_OPBITEN      (1 << 5)
+#define AD9833_DIV2         (1 << 3)
+#define AD9833_MODE         (1 << 1)
 
 
 
@@ -71,14 +75,14 @@ void AD9833::begin()
 void AD9833::reset()
 {
   hardwareReset();
-  _control = AD9833_B28;
+  _control = AD9833_B28;  //  implicit select sine wave.
   writeControlRegister(_control);
 }
 
 
 void AD9833::hardwareReset()
 {
-  writeControlRegister(_control | 0x0100);
+  writeControlRegister(_control | AD9833_RESET);
   //  reset all library variables to be in "sync" with hardware.
   _control  = 0;
   _waveform = AD9833_OFF;
@@ -148,7 +152,7 @@ float AD9833::setFrequency(float freq, uint8_t channel)
   //  rounding
   uint32_t fr = round(_freq[channel] * (268435456.0 / 25000000.0));
 
-  writeFreqRegister(channel, fr);
+  writeFrequencyRegister(channel, fr);
 
   return _freq[channel];
 }
@@ -233,7 +237,7 @@ bool AD9833::usesHWSPI()
 
 ////////////////////////////////////////////////////////////////
 //
-//  PRIVATE
+//  LOW LEVEL API - Expert users only
 //
 void AD9833::writeControlRegister(uint16_t value)
 {
@@ -242,35 +246,91 @@ void AD9833::writeControlRegister(uint16_t value)
 }
 
 
-void AD9833::writeFreqRegister(uint8_t reg, uint32_t freq)
+void AD9833:: writeFrequencyRegister(uint8_t channel, uint32_t freq)
 {
-  uint16_t data = 0;
-  if (reg > 1) return;
-  if (reg == 0) data = 0x4000;  //  bit 15 and 14    01
-  if (reg == 1) data = 0x8000;  //  bit 15 and 14    10
+  uint16_t LSB = 0;
+  uint16_t MSB = 0;
+  if (channel > 1) return;
+  if (channel == 0) LSB = AD9833_FREG0;  //  bit 15 and 14    01
+  if (channel == 1) LSB = AD9833_FREG1;  //  bit 15 and 14    10
+  //  copy channel mask.
+  MSB = LSB;
+
+  //  be sure B28 bit is set.
+  _control |= AD9833_B28;
+  writeControlRegister(_control);
 
   //  28 bits in two sets of 14
-  data |= (freq & 0x3FFF);  //  least significant 14 bits
-  writeData(data);
+  LSB |= (freq & 0x3FFF);
+  MSB |= ((freq >> 14) & 0x3FFF);
 
-  data &= 0xC000;           //  remove freq data LSB
-  data |= (freq >> 14);     //  most significant  14 bits
-  writeData(data);
+  //  faster to write them in one SPI transaction
+  writeData28(LSB, MSB);
+  //  first send the least significant 14 bits
+  //  writeData(LSB);
+  //  then send the most significant 14 bits
+  //  writeData(MSB);
 }
 
 
-void AD9833::writePhaseRegister(uint8_t reg, uint16_t value)
+void AD9833::writePhaseRegister(uint8_t channel, uint16_t value)
 {
   uint16_t data = 0;
-  if (reg > 1) return;
-  if (reg == 0) data = 0xC000;  //  bit 15 and 14 and 13   110
-  if (reg == 1) data = 0xE000;  //  bit 15 and 14 and 13   111
+  if (channel > 1) return;
+  if (channel == 0) data = 0xC000;  //  bit 15 and 14 and 13   110
+  if (channel == 1) data = 0xE000;  //  bit 15 and 14 and 13   111
 
   data |= (value & 0x0FFF);
   writeData(data);
 }
 
 
+
+///////////////////////////////////////////////////////////////////
+//
+//  EXPERIMENTAL
+//
+void AD9833::writeFrequencyRegisterLSB(uint8_t channel, uint16_t LSB)
+{
+  if (channel > 1) return;
+  //  force 14 bit
+  LSB &= 0x3FFF;
+  if (channel == 0) LSB |= AD9833_FREG0;  //  bit 15 and 14    01
+  if (channel == 1) LSB |= AD9833_FREG1;  //  bit 15 and 14    10
+
+  //  be sure B28 and HLB bit is cleared.
+  _control &= ~AD9833_B28;
+  _control &= ~AD9833_HLB;
+  writeControlRegister(_control);
+
+  //  send the least significant 14 bits
+  writeData(LSB);
+}
+
+
+void AD9833::writeFrequencyRegisterMSB(uint8_t channel, uint16_t MSB)
+{
+  if (channel > 1) return;
+  //  force 14 bit
+  MSB &= 0x3FFF;
+  if (channel == 0) MSB |= AD9833_FREG0;  //  bit 15 and 14    01
+  if (channel == 1) MSB |= AD9833_FREG1;  //  bit 15 and 14    10
+
+  //  be sure B28 is cleared and HLB bit is set.
+  _control &= ~AD9833_B28;
+  _control |= AD9833_HLB;
+  writeControlRegister(_control);
+
+  //  send the least significant 14 bits
+  writeData(MSB);
+}
+
+
+
+///////////////////////////////////////////////////////////////////
+//
+//  PRIVATE
+//
 void AD9833::writeData(uint16_t data)
 {
   if (_useSelect) digitalWrite(_selectPin, LOW);
@@ -296,6 +356,43 @@ void AD9833::writeData(uint16_t data)
   }
   if (_useSelect) digitalWrite(_selectPin, HIGH);
 }
+
+
+void AD9833::writeData28(uint16_t LSB, uint16_t MSB)
+{
+  if (_useSelect) digitalWrite(_selectPin, LOW);
+  if (_hwSPI)
+  {
+    _mySPI->beginTransaction(_spi_settings);
+    _mySPI->transfer16(LSB);
+    _mySPI->transfer16(MSB);
+    _mySPI->endTransaction();
+  }
+  else
+  {
+    //  local variables is fast.
+    uint8_t clk = _clockPin;
+    uint8_t dao = _dataPin;
+    //  MSBFIRST
+    for (uint16_t mask = 0x8000; mask; mask >>= 1)
+    {
+      digitalWrite(dao, (LSB & mask) !=0 ? HIGH : LOW);
+      digitalWrite(clk, LOW);
+      digitalWrite(clk, HIGH);
+    }
+    
+    for (uint16_t mask = 0x8000; mask; mask >>= 1)
+    {
+      digitalWrite(dao, (MSB & mask) !=0 ? HIGH : LOW);
+      digitalWrite(clk, LOW);
+      digitalWrite(clk, HIGH);
+    }
+    
+    digitalWrite(dao, LOW);
+  }
+  if (_useSelect) digitalWrite(_selectPin, HIGH);
+}
+
 
 
 //  -- END OF FILE --